refactor(hir): newtype parameter indexes (ParamIndex)
Wrap ParamId.index (u32) in a #[repr(transparent)] ParamIndex newtype.
Construction sites use ParamIndex::from_usize; the pattern-inference read
passes .as_u32() into param_ty (kept on u32 for the smaller diff). The
distinct local LatentComptimeParam.index stays usize. ParamIndex derives
salsa::Update.

+++ crates/hir/src/nameres/queries.rs
@@ -149,7 +149,7 @@ pub fn resolve_body_with_imports_and_policy<'db>(
     resolver.with_type_vars(&context.type_vars, |resolver| {
         resolver.with_scope(|resolver| {
             for (index, param) in context.params.iter().enumerate() {
-                resolver.add_param(body, index as u32, &param.name);
+                resolver.add_param(body, index, &param.name);
             }
             resolver.body(body);
         });
